+++ ft_memset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mhuescar < mhuescar@student.42malaga.co    +#+  +:+       +#+        */
+/*   By: mhuescar <mhuescar@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/13 12:31:39 by mhuescar          #+#    #+#             */
-/*   Updated: 2024/12/13 12:36:32 by mhuescar         ###   ########.fr       */
+/*   Updated: 2024/12/15 12:58:13 by mhuescar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
    #include <stdio.h>
 void *ft_memset(void *b, int c, size_t len)
 { 
-    int i;
+    size_t i;
     i = 0;
     unsigned char *ptr;
     ptr = (unsigned char*)b;
